requirements.txt
numpy == 1.21.5
pandas == 1.4.4
streamlit
scikit-learn
nltk
sastrawi
pip == 22.2.2
snowflake-snowpark-python
streamlit_option_menu
